Meu primeiro projeto intelliJ

+++ src/br/com/dio/calculadora/Calculadora.java
@@ -5,16 +5,16 @@ import java.util.Scanner;
 public class Calculadora {
     public static void main(String[] args) {
         Scanner scan = new Scanner(System.in);
-        float a,b;
+        int a,b;
         System.out.println("Digite o primeiro valor: ");
         a = scan.nextInt();
         System.out.println("Digite o segundo valor: ");
         b = scan.nextInt();
 
-       float somar = somar(a , b);
-       float subtrair = subtrair(a , b);
-       float multiplicar = multiplicar(a , b);
-       float dividir = dividir(a , b);
+       int somar = somar(a , b);
+       int subtrair = subtrair(a , b);
+       int multiplicar = multiplicar(a , b);
+       int dividir = dividir(a , b);
 
         System.out.println("somar: " + somar);
         System.out.println("subtrair: " + subtrair);
@@ -22,16 +22,16 @@ public class Calculadora {
         System.out.println("dividir: " + dividir);
     }
 
-    public static float somar(float a, float b){
+    public static int somar(int a, int b){
         return a + b;
     }
-    public static float subtrair(float a, float b){
+    public static int subtrair(int a, int b){
         return a - b;
     }
-    public static float multiplicar(float a, float b){
+    public static int multiplicar(int a, int b){
         return a * b;
     }
-    public static float dividir(float a, float b){
+    public static int dividir(int a, int b){
         return a / b;
     }
